added roles

+++ server/DanceApp/Queries/DanceMutation.cs
@@ -1,8 +1,10 @@
 ﻿using DanceApp.Exceptions;
+using DanceApp.Extensions;
 using DanceApp.Services;
 using Dataprovider.Exceptions;
 using Dataprovider.Models;
 using Dataprovider.Repositories;
+using System.Security.Claims;
 
 namespace DanceApp.Queries;
 
@@ -10,14 +12,21 @@ namespace DanceApp.Queries;
 public class DanceMutation
 {
     [Error<BlobException>]
+    [Error<UnauthorizedException>]
     public async Task<Dance> AddDance(
         [Service] DanceRepository repository,
         [Service] BlobService blobService,
+        ClaimsPrincipal claims,
         string name,
         string? imageBase64,
         string? imageUrl
         )
     {
+        var isAdmin = claims.IsAdmin();
+        if (!isAdmin)
+        {
+            throw new UnauthorizedException("User is not authorized to add dances");
+        }
         if (!string.IsNullOrEmpty(imageBase64))
         {
             var blob = await blobService.Upload($"Dance_{name}", imageBase64);
@@ -34,15 +43,22 @@ public class DanceMutation
     }
 
     [Error<BlobException>]
+    [Error<UnauthorizedException>]
     public async Task<Dance> UpdateDance(
         [Service] DanceRepository repository,
         [Service] BlobService blobService,
+        ClaimsPrincipal claims,
         [ID] int id,
         string name,
         string? imageBase64,
         string? imageUrl
         )
     {
+        var isAdmin = claims.IsAdmin();
+        if (!isAdmin)
+        {
+            throw new UnauthorizedException("User is not authorized to update dances");
+        }
         var danceToUpdate = repository.GetDanceById(id);
         bool imageChanged = !string.IsNullOrEmpty(imageBase64) || imageUrl != danceToUpdate.ImageUrl;
         var previousImageUrl = danceToUpdate.ImageUrl;
@@ -69,12 +85,19 @@ public class DanceMutation
 
     [Error<NotFoundException>]
     [Error<BlobException>]
+    [Error<UnauthorizedException>]
     public async Task<Dance> DeleteDance(
         [Service] DanceRepository repository,
         [Service] BlobService blobService,
+        ClaimsPrincipal claims,
         [ID] int danceId
         )
     {
+        var isAdmin = claims.IsAdmin();
+        if (!isAdmin)
+        {
+            throw new UnauthorizedException("User is not authorized to delete dances");
+        }
         var dance = repository.DeleteDance(danceId);
         if (!string.IsNullOrEmpty(dance?.ImageUrl))
         {
